+++ tests/ClientTest.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Sentry\Tests;
 
+use PHPUnit\Framework\MockObject\Matcher\Invocation;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 use Sentry\ClientBuilder;
 use Sentry\Event;
 use Sentry\Options;
@@ -13,6 +15,7 @@ use Sentry\SentrySdk;
 use Sentry\Serializer\Serializer;
 use Sentry\Severity;
 use Sentry\Stacktrace;
+use Sentry\State\Scope;
 use Sentry\Transport\TransportFactoryInterface;
 use Sentry\Transport\TransportInterface;
 
@@ -309,31 +312,91 @@ class ClientTest extends TestCase
     }
 
     /**
-     * @dataProvider sampleRateAbsoluteDataProvider
+     * @dataProvider processEventDiscardsEventWhenItIsSampledDueToSampleRateOptionDataProvider
      */
-    public function testSampleRateAbsolute(float $sampleRate): void
+    public function testProcessEventDiscardsEventWhenItIsSampledDueToSampleRateOption(float $sampleRate, Invocation $transportCallInvocationMatcher, Invocation $loggerCallInvocationMatcher): void
     {
+        /** @var TransportInterface&MockObject $transport */
         $transport = $this->createMock(TransportInterface::class);
-        $transport->expects(0 == $sampleRate ? $this->never() : $this->exactly(10))
+        $transport->expects($transportCallInvocationMatcher)
             ->method('send');
 
-        $transportFactory = $this->createTransportFactory($transport);
+        /** @var LoggerInterface&MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($loggerCallInvocationMatcher)
+            ->method('info')
+            ->with('The event will be discarded because it has been sampled.', $this->callback(static function (array $context): bool {
+                return isset($context['event']) && $context['event'] instanceof Event;
+            }));
 
-        $client = (new ClientBuilder(new Options(['sample_rate' => $sampleRate])))
-            ->setTransportFactory($transportFactory)
+        $client = ClientBuilder::create(['sample_rate' => $sampleRate])
+            ->setTransportFactory($this->createTransportFactory($transport))
+            ->setLogger($logger)
             ->getClient();
 
         for ($i = 0; $i < 10; ++$i) {
-            $client->captureMessage('foobar');
+            $client->captureMessage('foo');
         }
     }
 
-    public function sampleRateAbsoluteDataProvider(): array
+    public function processEventDiscardsEventWhenItIsSampledDueToSampleRateOptionDataProvider(): \Generator
     {
-        return [
-            'sample rate 0' => [0],
-            'sample rate 1' => [1],
+        yield [
+            0,
+            $this->never(),
+            $this->exactly(10),
         ];
+
+        yield [
+            1,
+            $this->exactly(10),
+            $this->never(),
+        ];
+    }
+
+    public function testProcessEventDiscardsEventWhenBeforeSendCallbackReturnsNull(): void
+    {
+        /** @var LoggerInterface&MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->once())
+            ->method('info')
+            ->with('The event will be discarded because the "before_send" callback returned `null`.', $this->callback(static function (array $context): bool {
+                return isset($context['event']) && $context['event'] instanceof Event;
+            }));
+
+        $options = [
+            'before_send' => static function () {
+                return null;
+            },
+        ];
+
+        $client = ClientBuilder::create($options)
+            ->setLogger($logger)
+            ->getClient();
+
+        $client->captureMessage('foo');
+    }
+
+    public function testProcessEventDiscardsEventWhenEventProcessorReturnsNull(): void
+    {
+        /** @var LoggerInterface&MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->once())
+            ->method('info')
+            ->with('The event will be discarded because one of the event processors returned `null`.', $this->callback(static function (array $context): bool {
+                return isset($context['event']) && $context['event'] instanceof Event;
+            }));
+
+        $client = ClientBuilder::create([])
+            ->setLogger($logger)
+            ->getClient();
+
+        $scope = new Scope();
+        $scope->addEventProcessor(static function () {
+            return null;
+        });
+
+        $client->captureMessage('foo', Severity::debug(), $scope);
     }
 
     /**