+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enforce a timeout for connecting to the server and for the requests instead of waiting indefinitely (#979)
 - Add `RequestFetcherInterface` to allow customizing the request data attached to the logged event (#984)
+- Log internal debug and error messages to a PSR-3 compatible logger (#989)
 
 ## 2.3.2 (2020-03-06)